feat: Implement clarification session management with memory functions
